completed part4

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,34 +1,59 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const { title, author, url, likes } = request.body
 
-  // 4.12: title and url are required
-  if (!title || !url) {
-    return response.status(400).json({
-      error: 'title or url missing',
-    })
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
 
-  // 4.11: likes defaults to 0
+  if (!title || !url) {
+    return response.status(400).json({ error: 'title or url missing' })
+  }
+
   const blog = new Blog({
     title,
     author,
     url,
     likes: likes || 0,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
+
+  response.json(blogs)
+})
+
+
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) return response.status(404).end()
+
+  if (!request.user || blog.user.toString() !== request.user._id.toString()) {
+    return response.status(401).json({ error: 'only creator can delete blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
